Iteration 2 complete
I kept with my assumption that special characters are valid when the other requirements are met in this iteration. Because of that, entering 8 or more special characters without any numbers or letters will falsely trigger the not min length error, but adding one of each will pass it. I will move forward with this being a known issue.