iteration3 complete
I also fixed my counter in the whyWeak function I created. In regards to my failure to see the obvious when it comes to the min length, I have no words. I am laughing and trying to get over it.  I was too focused on regex when there were simpler options.